requirements.txt
google-generativeai==0.4.1
pillow==10.3.0
streamlit==1.31.1
streamlit-option-menu==0.3.13

